+++ rebo/src/parser/scope.rs
@@ -1,7 +1,7 @@
 use indexmap::map::IndexMap;
 use crate::parser::{Binding, Generic, ExprLabelDef};
 use std::fmt::{self, Formatter, Display};
-use std::sync::atomic::{AtomicU32, Ordering};
+use std::sync::atomic::{AtomicU64, Ordering};
 
 pub struct Scope<'i> {
     pub idents: IndexMap<String, Binding<'i>>,
@@ -25,20 +25,20 @@ pub enum ScopeType<'i> {
 }
 
 #[derive(Debug, Clone, Copy, Ord, PartialOrd, Eq, PartialEq, Hash)]
-pub struct BindingId(u32);
+pub struct BindingId(u64);
 impl Display for BindingId {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.0)
     }
 }
 
-static NEXT_BINDING_ID: AtomicU32 = AtomicU32::new(0);
+static NEXT_BINDING_ID: AtomicU64 = AtomicU64::new(0);
 
 impl BindingId {
     pub fn unique() -> BindingId {
         let id = NEXT_BINDING_ID.fetch_add(1, Ordering::SeqCst);
         // why do we even have this check?!
-        if id == u32::MAX {
+        if id == u64::MAX {
             panic!("binding id overflow");
         }
         BindingId(id)
